Change ReplacementInviteesTest to use MockedEventList
